refactor(app-builder): move storage libs to /shared/lib folder
