+++ apps/app-builder/src/app/model/useSetup.ts
@@ -1,5 +1,5 @@
 import { textPreferencesStatePersist } from '@/entities/textPreferences'
-import { LocalStorage } from '@/shared/storage'
+import { LocalStorage } from '@/shared/lib/storage'
 
 export function useSetup() {
   const storage = chrome.extension ? chrome.storage.local : new LocalStorage()
