+++ apps/app-builder/src/entities/textPreferences/storage/textPreferencesStatePersist.ts
@@ -1,5 +1,5 @@
 import { watchEffect } from 'vue'
-import { Storage } from '@/shared/storage'
+import { Storage } from '@/shared/lib/storage'
 import { TextPreferences } from '../model/TextPreferences'
 import { TEXT_PREFERENCES_STORAGE_KEY } from '../config/storage'
 import { useTextPreferences } from '../model/state/useTextPreferences'
