+++ apps/app-builder/src/shared/lib/storage/index.ts
@@ -1,4 +1,3 @@
-// TODO: move to lib/ folder
 export type { Storage, StorageItems } from './storage'
 export { MemoryStorage } from './memoryStorage'
 export { LocalStorage } from './localStorage'
